ClangTidy

+++ include/dd/MDDPackage.hpp
@@ -1476,7 +1476,7 @@ namespace dd {
             // NOLINTNEXTLINE(cppcoreguidelines-avoid-do-while)
             do {
                 ComplexNumbers::mul(tempCompNumb, tempCompNumb, currentEdge.weight);
-                std::size_t tmp = pathElements.at(currentEdge.nextNode->varIndx) - '0';
+                std::size_t tmp = pathElements.at(static_cast<std::size_t>(currentEdge.nextNode->varIndx)) - '0';
                 assert(tmp <= currentEdge.nextNode->edges.size());
                 currentEdge = currentEdge.nextNode->edges.at(tmp);
             } while (!currentEdge.isTerminal());
